+++ src/app/core/services/gen-id.generator.service.ts
@@ -1,0 +1,6 @@
+let counter = 0;
+
+export const genID = (): number => {
+  counter += 1;
+  return counter;
+}