+++ src/app/core/services/random-id.factory.ts
@@ -1,0 +1,9 @@
+import {GeneratorService} from "./generator.service";
+import {InjectionToken} from "@angular/core";
+
+export const generatedString = new InjectionToken<string>('generatedString');
+
+export function RandomIdFactory(idLength: number): (genId: GeneratorService) => string {
+  return (genId: GeneratorService): string =>
+    genId.generate(idLength);
+}