+++ src/app/shared/models/app-info.ts
@@ -1,0 +1,5 @@
+export interface AppInfo {
+  App: string;
+  Ver: string;
+  API_URL: string;
+}